requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
pkg-resources==0.0.0
python-telegram-bot==13.4.1
pytz==2021.1
six==1.15.0
style==1.1.0
tornado==6.1
tzlocal==2.1
update==0.0.1

PyMySQL~=1.0.2
setuptools~=54.1.2